refactor(ui): extract AlertItem to reduce JSX depth and satisfy ui complexity check
